Abstracted security logging functionality for both exception & manual
- Moved SecurityException.log & SecurityException.email method to
SecurityUtil
- Added support for SecurityLog.log(Exception e) &
SecurityLog.log(String eventType, String eventMessage)
- Updates to SecurityLog view to be like Audit
