+++ skyve-war/src/main/java/modules/admin/SecurityLog/views/edit.xml
@@ -3,76 +3,51 @@
 	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" name="edit"
 	title="{bizKey}"
 	xsi:schemaLocation="http://www.skyve.org/xml/view ../../../../schemas/view.xsd">
-	<form border="true" borderTitle="Security Log Details"
-		disabled="true">
-		<column responsiveWidth="3" />
-		<column responsiveWidth="6" />
-		<column />
-		<row>
-			<item align="right">
-				<label value="{disp:timestamp} *:" />
-			</item>
-			<item showLabel="false">
-				<textField binding="timestamp" />
-			</item>
+	<form border="true" borderTitle="Security Log Details" disabled="true">
+        <column responsiveWidth="3" />
+        <column responsiveWidth="6" />
+        <column />
+        <row>
+            <item showLabel="true" label="Timestamp">
+            	<blurb>{timestamp}</blurb>
+            </item>
+		</row>
+        <row>
+            <item showLabel="true" label="Thread ID">
+            	<blurb>{threadID}</blurb>
+            </item>
 		</row>
 		<row>
-			<item align="right">
-				<label value="{disp:threadID} *:" />
-			</item>
-			<item showLabel="false">
-				<textField binding="threadID" />
-			</item>
+            <item showLabel="true" label="Source IP">
+            	<blurb>{sourceIP}</blurb>
+            </item>
+        </row>
+        <row>
+            <item showLabel="true" label="Username">
+            	<blurb>{username}</blurb>
+            </item>
 		</row>
 		<row>
-			<item align="right">
-				<label value="{disp:sourceIP} *:" />
-			</item>
-			<item showLabel="false">
-				<textField binding="sourceIP" />
-			</item>
-		</row>
-		<row>
-			<item align="right">
-				<label value="{disp:username} *:" />
-			</item>
-			<item showLabel="false">
-				<textField binding="username" />
-			</item>
-		</row>
-		<row>
-			<item align="right">
-				<label value="{disp:loggedInUserId} *:" />
-			</item>
-			<item showLabel="false">
-				<textField binding="loggedInUserId" />
-			</item>
-		</row>
-		<row>
-			<item align="right">
-				<label value="{disp:exceptionType} *:" />
-			</item>
-			<item showLabel="false">
-				<combo binding="exceptionType" />
-			</item>
-		</row>
-		<row>
-			<item align="right">
-				<label value="{disp:exceptionMessage} *:" />
-			</item>
-			<item showLabel="false">
-				<textField binding="exceptionMessage" />
-			</item>
-		</row>
-		<row>
-			<item align="right">
-				<label value="{disp:provenance} *:" />
-			</item>
-			<item showLabel="false">
-				<textField binding="provenance" />
-			</item>
-		</row>
-	</form>
+            <item showLabel="true" label="Logged in User ID">
+            	<blurb>{loggedInUserId}</blurb>
+            </item>
+        </row>
+        <row>
+            <item showLabel="true" label="Event Type">
+            	<blurb>{eventType}</blurb>
+            </item>
+        </row>
+         <row>
+            <item showLabel="true" label="Event Message">
+            	<blurb>{eventMessage}</blurb>
+            </item>
+        </row>
+         <row>
+            <item showLabel="true" label="Provenance">
+            	<blurb>{provenance}</blurb>
+            </item>
+        </row>
+    </form>
 	<actions>
 		<cancel />
 	</actions>
